refactor: change prisma & db connection to another folder
