+++ package.json
@@ -30,20 +30,20 @@
     "next": "13.1.4",
     "next-auth": "^4.18.8",
     "npm": "^9.4.1",
-
     "pdfmake": "^0.2.7",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "rsuite": "^5.28.1",
-
     "react-hook-form": "^7.43.0",
+    "rsuite": "^5.28.2",
     "typescript": "4.9.4",
+    "yup": "^1.0.2",
     "zod": "^3.20.2"
   },
   "devDependencies": {
-    "@types/pdfmake": "^0.2.2",
     "@types/bcrypt": "^5.0.0",
     "@types/node": "^18.11.18",
+    "@types/pdfmake": "^0.2.2",
+    "@types/yup": "^0.32.0",
     "@typescript-eslint/eslint-plugin": "^5.13.0",
     "@typescript-eslint/parser": "^5.0.0",
     "autoprefixer": "^10.4.13",
